AI Plant Guardian/scripts/requirements.txt
streamlit
tensorflow
opencv-python
Pillow
numpy
matplotlib
